+++ src/main/java/com/project/demo/domain/order/service/OrderService.java
@@ -11,11 +11,11 @@ public interface OrderService {
 
     public String buyingStock(Long userId, String ticker, int quantity);
 
-    public void executeBuy(Order order, User user, Stock stock, int price, int quantity, int totalPrice);
+    public void executeBuy(Order order, User user, Stock stock, int price, int quantity, long totalPrice);
 
     public String sellingStock(Long userId, String ticker, int quantity);
 
-    public void executeSell(Order order, User user, Stock stock, int price, int quantity, int totalPrice);
+    public void executeSell(Order order, User user, Stock stock, int price, int quantity, long totalPrice);
 
     public String reserveBuy(Long userId, String ticker, int quantity, int targetPrice);
 
@@ -25,7 +25,8 @@ public interface OrderService {
 
     /**
      * 특정 종목의 예약 주문 체결 (이벤트 기반)
-     * @param ticker 종목 코드
+     * 
+     * @param ticker       종목 코드
      * @param currentPrice 현재가
      */
     public void executeReservedOrdersForTicker(String ticker, int currentPrice);
@@ -37,8 +38,5 @@ public interface OrderService {
     public List<OrderResponse> getNormalOrders(Long userId);
 
     public List<OrderResponse> getReservationOrders(Long userId);
-
-
-
 
 }
